class MissionComponent is now renamed MissionAdvancedRun and inherits from MissionRun.
